requirements.txt
streamlit>=1.23.0
openai>=0.27.2
python-dotenv>=1.0.0
pytube>=15.0.0
tiktoken>=0.3.3
langchain>=0.0.181
youtube-transcript-api==0.6.0
azure-cosmosdb-table==1.0.6
pandas==2.0.1